fixed writing of colour buffer to include the i2c register address as well

+++ include/neopixel.h
@@ -27,7 +27,7 @@ public:
 
 	int 	SetPin			(int input);
 	int 	SetLength	 	(int input);
-	int 	SetColours 		(uint8_t *buf);
+	int 	SetColours 		(int *buf);
